Panel update and Helix changes
What used to be Segment list is now helix as it makes more sense. props,
that are mean for whole helix are in helix. and props of segments are in
segments + methods transfered.
Panel are now defined by template name type is not important anymore.
But also has to be specified if unique. If yes it means that if same
template panel is already created the existing is just updated not
removed and respawned.
So now its like HELIX - Segment - Mesh.